requirements.txt
PyYAML>=6.0
jsonschema>=4.0.0
requests>=2.25.0
